requirements/base.txt
packaging==17.1
six==1.11.0
appdirs==1.4.3
pylint==1.8.4
pep8==1.7.1
coverage==4.5.1
psycopg2==2.7.4
djangorestframework==3.8.2
Markdown==2.6.11
django-rest-swagger==2.2.0
django-cors-headers==2.2.0
django-filter==1.1.0
Django==2.0.5
boto==2.48.0
django-ses==0.8.5
pika==0.11.2
python-docx==0.8.6
djangorestframework-filters==0.10.2
djangorestframework-bulk==0.2.1
celery==4.1.0
suds-py3==1.3.3.0
xmltodict==0.11.0
djangorestframework-xml==1.3.0
boto3==1.7.12
inflection==0.3.1
python-stdnum==1.9
Faker==0.8.13
zxcvbn==4.4.25
zxcvbn-python==4.4.24
django-cors-middleware==1.3.1
psycopg2-binary==2.7.4
django-model-utils==3.1.1
Pillow==5.1.0
django-imagekit==4.0.2
